wip: Remove deprecated content and update import paths
