+++ src/core/index.ts
@@ -16,10 +16,16 @@ const args = _args;
 
     const hook = (win: Window & typeof globalThis, data: WindowStorage | undefined) => {
       if (!data) return;
-      if (!existParentDomain(storage.blacklist, data.host) && existParentDomain(storage.whitelist, data.host)) return;
-      if (win.location.hostname !== data.host
-        && !existParentDomain(storage.blacklist, win.location.hostname)
-        && existParentDomain(storage.whitelist, win.location.hostname)) return;
+
+      const ps = storage.policies;
+      if (ps.isBlacklistMode) {
+        if (!existParentDomain(ps.blacklist, data.host)) return;
+        if (win.location.hostname !== data.host && !existParentDomain(ps.blacklist, win.location.hostname)) return;
+      } else {
+        if (existParentDomain(ps.whitelist, data.host)) return;
+        if (win.location.hostname !== data.host && existParentDomain(ps.whitelist, win.location.hostname)) return;
+      }
+
       try {
         FingerprintContext.hookWindow(win, {
           info: data,
