+++ src/popup/stores/storage.ts
@@ -1,249 +1,92 @@
-import { deepProxy, selectChildDomains, tryUrl } from "@/utils/base"
+import i18n from "@/locales"
 import { sendToBackground } from "@/utils/message"
-import { debounce, sharedAsync } from "@/utils/timer"
+import { debounce } from "@/utils/timer"
 import { create } from "zustand"
 
 type State = {
-  storage?: LocalStorage
-  config?: LocalStorageConfig  // Proxy
-  whitelist?: string[]
-  blacklist?: string[]
   version: number
+  storage?: LocalStorage
+  config?: LocalStorageConfig
+  policies?: LocalStoragePolicies
 }
 
 type Actions = {
-  syncLoadStorage: (storage: LocalStorage) => void
-  loadStorage: () => Promise<void>
-  importStorage: (ss: DeepPartial<LocalStorage>) => Promise<void>
+  loadStorage: () => void
+  importStorage: (data: DeepPartial<LocalStorage>) => Promise<void>
   saveConfig: () => void
-  addWhitelist: (list: string | string[]) => void
-  deleteWhitelist: (list: string | string[]) => void
-  cleanWhitelist: () => void
-  addBlacklist: (list: string | string[]) => void
-  deleteBlacklist: (list: string | string[]) => void
-  cleanBlacklist: () => void
+  savePolicies: () => void
 }
 
 export const useStorageStore = create<State & Actions>(((set, get) => {
 
-  const saveConfig = debounce(() => {
-    const config = get().storage?.config
-    if (config) {
+  const loadStorage = async () => {
+    const s = await chrome.storage.local.get() as LocalStorage
+    set({
+      storage: s,
+      config: s.config,
+      policies: s.policies,
+    })
+  }
+
+  const importStorage = async (data: DeepPartial<LocalStorage>) => {
+    const res = await sendToBackground({
+      type: 'storage.import',
+      storage: data,
+    })
+
+    if (res.ok && res.storage) {
+      set({
+        storage: res.storage,
+        config: res.storage.config,
+        policies: res.storage.policies,
+      })
+    } else {
+      throw new Error(
+        (res.messageKey ? i18n.t(res.messageKey) : res.message) ?? 'Import failed'
+      )
+    }
+  }
+
+  const _saveConfig = debounce(async () => {
+    const v = get().config;
+    if (v) {
       sendToBackground({
         type: 'config.set',
-        config,
+        config: v,
       })
     }
   }, 200)
 
-  const proxyConfig = (config: LocalStorageConfig) => {
-    return deepProxy(config, {
-      set(target, key, value, receiver) {
-        const res = Reflect.set(target, key, value, receiver);
-        saveConfig()
-        return res;
-      },
-    });
-  }
-
-  const syncLoadStorage = (storage: LocalStorage) => {
-    set({
-      storage,
-      config: proxyConfig(storage.config),
-      whitelist: storage.whitelist,
-      blacklist: storage.blacklist,
-      version: get().version + 1,
-    })
-  }
-
-  const loadStorage = sharedAsync(async () => {
-    syncLoadStorage(await chrome.storage.local.get() as LocalStorage)
-  })
-
-  const importStorage = async (ss: DeepPartial<LocalStorage>) => {
-    const state = get()
-    const storage = state.storage
-    if (!storage) throw 'tip.err.config-unloaded';
-
-    let isImported = false;
-
-    /* 导入配置 */
-    if (ss.config) {
-      // 排除一些不支持导入的配置
-      delete ss.config.prefs;
-      delete ss.config.action?.fastInject;
-      // 导入
-      const _config = await sendToBackground({
-        type: 'config.set',
-        config: ss.config,
-        result: true,
+  const _savePolicies = debounce(async () => {
+    const v = get().policies;
+    if (v) {
+      sendToBackground({
+        type: 'policies.set',
+        policies: v,
       })
-      if (_config) {
-        storage.config = _config
-        set({
-          config: proxyConfig(storage.config),
-          version: state.version + 1,
-        })
-        isImported = true
-      }
     }
+  }, 200)
 
-    /* 导入白名单 */
-    if (ss.whitelist?.length) {
-      addWhitelist(ss.whitelist)
-      isImported = true
-    }
-
-    /* 导入黑名单 */
-    if (ss.blacklist?.length) {
-      addBlacklist(ss.blacklist)
-      isImported = true
-    }
-
-    /* 导入为空 */
-    if (!isImported) {
-      throw 'Content is empty'
-    }
+  const saveConfig = () => {
+    _saveConfig()
+    set(({ version }) => ({
+      version: version + 1,
+    }))
   }
 
-  const addWhitelist = (list: string | string[]) => {
-    const storage = get().storage
-    if (!storage) return;
-    if (!Array.isArray(list)) list = [list];
-    const whitelist = [...storage.whitelist]
-    const del: string[] = []
-    for (const item of list) {
-      const url = tryUrl('http://' + item)
-      if (!url) continue;
-      for (const domain of selectChildDomains(whitelist, url.hostname)) {
-        const index = whitelist.indexOf(domain)
-        if (index !== -1) {
-          whitelist.splice(index, 1)
-          del.push(domain)
-        }
-      }
-      whitelist.push(url.hostname)
-    }
-    storage.whitelist = whitelist
-    set({ whitelist: storage.whitelist })
-    sendToBackground({
-      type: 'whitelist.update',
-      data: {
-        add: list,
-        del,
-      },
-    })
-  }
-
-  const addBlacklist = (list: string | string[]) => {
-    const storage = get().storage
-    if (!storage) return;
-    if (!Array.isArray(list)) list = [list];
-    const blacklist = [...storage.blacklist]
-    const del: string[] = []
-    for (const item of list) {
-      const url = tryUrl('http://' + item)
-      if (!url) continue;
-      for (const domain of selectChildDomains(blacklist, url.hostname)) {
-        const index = blacklist.indexOf(domain)
-        if (index !== -1) {
-          blacklist.splice(index, 1)
-          del.push(domain)
-        }
-      }
-      blacklist.push(url.hostname)
-    }
-    storage.blacklist = blacklist
-    set({ blacklist: storage.blacklist })
-    sendToBackground({
-      type: 'blacklist.update',
-      data: {
-        add: list,
-        del,
-      },
-    })
-  }
-
-  const deleteWhitelist = (list: string | string[]) => {
-    const storage = get().storage
-    if (!storage) return;
-    if (!Array.isArray(list)) list = [list];
-    const whitelist = [...storage.whitelist]
-    for (const item of list) {
-      const url = tryUrl('http://' + item)
-      if (!url) continue;
-      const index = whitelist.indexOf(url.hostname)
-      index !== -1 && whitelist.splice(index, 1)
-    }
-    storage.whitelist = whitelist
-    set({ whitelist: storage.whitelist })
-    sendToBackground({
-      type: 'whitelist.update',
-      data: {
-        del: list,
-      },
-    })
-  }
-
-  const deleteBlacklist = (list: string | string[]) => {
-    const storage = get().storage
-    if (!storage) return;
-    if (!Array.isArray(list)) list = [list];
-    const blacklist = [...storage.blacklist]
-    for (const item of list) {
-      const url = tryUrl('http://' + item)
-      if (!url) continue;
-      const index = blacklist.indexOf(url.hostname)
-      index !== -1 && blacklist.splice(index, 1)
-    }
-    storage.blacklist = blacklist
-    set({ blacklist: storage.blacklist })
-    sendToBackground({
-      type: 'blacklist.update',
-      data: {
-        del: list,
-      },
-    })
-  }
-
-  const cleanWhitelist = () => {
-    const storage = get().storage
-    if (!storage) return;
-    storage.whitelist = []
-    set({ whitelist: storage.whitelist })
-    sendToBackground({
-      type: 'whitelist.update',
-      clean: true,
-    })
-  }
-
-  const cleanBlacklist = () => {
-    const storage = get().storage
-    if (!storage) return;
-    storage.blacklist = []
-    set({ blacklist: storage.blacklist })
-    sendToBackground({
-      type: 'blacklist.update',
-      clean: true,
-    })
+  const savePolicies = () => {
+    _savePolicies()
+    set(({ version }) => ({
+      version: version + 1,
+    }))
   }
 
   return {
-    storage: undefined,
-    config: undefined,
-    whitelist: undefined,
-    blacklist: undefined,
     version: 0,
 
-    syncLoadStorage,
     loadStorage,
     importStorage,
     saveConfig,
-    addWhitelist,
-    deleteWhitelist,
-    cleanWhitelist,
-    addBlacklist,
-    deleteBlacklist,
-    cleanBlacklist,
+    savePolicies,
   }
 }))
